feat(front): Patient views migration

+++ client-app/src/main/java/com/piedraazul/app_client/services/UserService.java
@@ -7,11 +7,11 @@ public interface UserService {
 
     boolean regUser(User newUser);
 
-    RoleUserEnum authUser(int cedUser, String password);
+    RoleUserEnum authUser(Long cedUser, String password);
 
     boolean recoverPassword(int cedUser, String answer);
 
     boolean updatePassword(int cedUser, String newPassword);
     
-    User findByCedUser(int cedUser);
+    User findByCedUser(Long cedUser);
 }
